refactor(index): move getStampOptions to utils and update ESLint ignores
